+++ GenForm-NPM/React/README.md
@@ -1,0 +1,10 @@
+# genform for react
+
+Genform is a JavaScript library for creating forms. It is designed to be simple and flexible. You can create forms with a few lines of code. It is also possible to create complex forms with a lot of customization.
+
+GenForm for react allows you to use genform in your react project as a react component.
+
+## Dependencies
+
+- [React](https://www.npmjs.com/package/react) - React library
+- [GenForm - Core](https://www.npmjs.com/package/@genform/core) - Core library
