+++ GenForm-NPM/React/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@genform/react",
+  "version": "0.1.2",
+  "description": "A form generator using JS based on a JSON object for React projects",
+  "main": "index.js",
+  "dependencies": {
+    "@genform/core": "^0.0.4",
+    "react": "^18.2.0"
+  },
+  "scripts": {
+    "test": "node index.js",
+    "lint": "eslint ../../.eslintrc.js"
+  },
+  "keywords": [
+    "form-generator",
+    "form",
+    "react"
+  ],
+  "author": "GenForm",
+  "license": "ISC",
+  "devDependencies": {
+    "eslint": "^8.55.0",
+    "prettier": "3.1.1"
+  }
+}
